Add OS exercise Starcraft in C
The homework description isn't that great...
There is some ambiguity in regards the command centers and how their minerals are shared/accumulated.
There is ambiguity in the total collected minerals that are to be displayed at the end.
Should purchases be included or excluded in the final value?

Even though the idea was to teach shared memory synchronization, I don't think it was a great homework.